+++ onedal/_config.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ==============================================================================
 
-"""Tools to expose sklearnex's config settings to daal4py level."""
+"""Tools to expose some sklearnex's config settings to onedal4py level."""
 
 import threading
 
@@ -26,12 +26,14 @@ _default_global_config = {
 _threadlocal = threading.local()
 
 
-def _get_daal4py_threadlocal_config():
-    if not hasattr(_threadlocal, "d4p_global_config"):
+def _get_onedal_threadlocal_config():
+    if not hasattr(_threadlocal, "global_config"):
         _threadlocal.d4p_global_config = _default_global_config.copy()
     return _threadlocal.d4p_global_config
 
 
+# TODO:
+# docstrings
 def _get_config():
     """Retrieve current values for configuration set by :func:`set_config`
     Returns
@@ -42,31 +44,5 @@ def _get_config():
     --------
     _set_config : Set global configuration.
     """
-    daal4py_config = _get_daal4py_threadlocal_config().copy()
-    return {**daal4py_config}
-
-
-def _set_config(target_offload=None, allow_fallback_to_host=None):
-    """Set global configuration
-    Parameters
-    ----------
-    target_offload : string or dpctl.SyclQueue, default=None
-        The device primarily used to perform computations.
-        If string, expected to be "auto" (the execution context
-        is deduced from input data location),
-        or SYCL* filter selector string. Global default: "auto".
-    allow_fallback_to_host : bool, default=None
-        If True, allows to fallback computation to host device
-        in case particular estimator does not support the selected one.
-        Global default: False.
-    See Also
-    --------
-    _get_config : Retrieve current values of the global configuration.
-    """
-
-    local_config = _get_daal4py_threadlocal_config()
-
-    if target_offload is not None:
-        local_config["target_offload"] = target_offload
-    if allow_fallback_to_host is not None:
-        local_config["allow_fallback_to_host"] = allow_fallback_to_host
+    onedal_config = _get_onedal_threadlocal_config().copy()
+    return {**onedal_config}
