+++ onedal/common/_policy.py
@@ -18,10 +18,6 @@ import sys
 
 from onedal import _backend, _is_dpc_backend
 
-oneapi_is_available = "daal4py.oneapi" in sys.modules
-if oneapi_is_available:
-    from daal4py.oneapi import _get_sycl_ctxt
-
 
 def _get_policy(queue, *data):
     data_queue = _get_queue(*data)
@@ -46,21 +42,9 @@ def _get_queue(*data):
     return None
 
 
-class _Daal4PyContextReset:
-    def __init__(self):
-        self._d4p_context = None
-        if oneapi_is_available:
-            self._d4p_context = _get_sycl_ctxt()
-
-    def __del__(self):
-        if self._d4p_context:
-            self._d4p_context.apply()
-
-
 class _HostInteropPolicy(_backend.host_policy):
     def __init__(self):
         super().__init__()
-        self._d4p_interop = _Daal4PyContextReset()
 
 
 if _is_dpc_backend:
@@ -68,8 +52,7 @@ if _is_dpc_backend:
     class _DataParallelInteropPolicy(_backend.data_parallel_policy):
         def __init__(self, queue):
             self._queue = queue
-            self._d4p_interop = _Daal4PyContextReset()
-            from daal4py.sklearn._device_offload import DummySyclQueue
+            from onedal._device_offload import DummySyclQueue
 
             if isinstance(queue, DummySyclQueue):
                 super().__init__(self._queue.sycl_device.get_filter_string())
