+++ sklearnex/tests/test_config.py
@@ -16,7 +16,7 @@
 
 import sklearn
 
-import daal4py
+import onedal
 import sklearnex
 
 
@@ -34,10 +34,10 @@ def test_set_config_works():
     )
 
     config = sklearnex.get_config()
-    config_d4p = daal4py.sklearn._get_config()
+    onedal_config = onedal._get_config()
     assert config["target_offload"] == "cpu:0"
     assert config["allow_fallback_to_host"]
     assert config["assume_finite"]
-    assert config_d4p["target_offload"] == "cpu:0"
-    assert config_d4p["allow_fallback_to_host"]
+    assert onedal_config["target_offload"] == "cpu:0"
+    assert onedal_config["allow_fallback_to_host"]
     sklearnex.set_config(**default_config)
